# apps/backend/openapi.yaml
openapi: 3.0.3
info:
  title: BigIn Asset API
  description: >-
    API quản lý tài sản BigIn Asset. Bao gồm xác thực (F01), quản lý tài sản (F02),
    phiếu mượn (F03), duyệt & giữ chỗ (F04), bàn giao/hoàn trả/lịch sử mượn (F05),
    sự cố & sửa chữa (F06), thông báo (F07) và quản trị người dùng/phân quyền (F08).

    Quy ước chung:
    - Envelope thành công: `{ "data": ... }`.
    - Envelope lỗi: `{ "error": string, "details"?: object }`.
    - Danh sách: `page=1`, `pageSize=20` (tối đa 100); page DTO `{ items, page, pageSize, total }`.
    - Danh sách sắp xếp mới nhất trước; Review Queue sắp xếp cũ nhất trước.
    - `expectedReturnDate` chỉ ngày `YYYY-MM-DD` theo múi giờ `Asia/Ho_Chi_Minh`, phải từ hôm nay trở đi.
    - Authorization theo effective permission code (không kiểm tra tên role).
    - Lỗi dự kiến: `400` input sai, `401` chưa xác thực, `403` thiếu quyền, `404` không tìm thấy, `409` xung đột trạng thái/dữ liệu.
  version: 1.0.0
  contact:
    name: BigIn SE Team
servers:
  - url: http://localhost:3000/api
    description: Local development server
...
paths:
...
  /assets/{id}/report-damaged:
    post:
      tags:
        - F06 - Asset Issues & Repair
      summary: Báo sự cố tài sản
      description: >-
        Tạo issue `REPORTED` mà không thay đổi trạng thái asset. Chỉ cần xác thực (không cần
        quyền tĩnh) vì người đang mượn cũng đủ điều kiện. Tên path giữ nguyên vì lý do tương thích.
      operationId: reportDamaged
      parameters:
        - $ref: '#/components/parameters/AssetId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReportIssueInput'
      responses:
        '201':
          description: Tạo báo cáo sự cố thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/AssetIssue'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          description: Người dùng đã xác thực nhưng không đủ điều kiện báo cáo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Tài sản không tồn tại
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    AssetId:
      name: id
      in: path
      required: true
      description: ID tài sản
      schema:
        type: integer
        minimum: 1
...
  responses:
    BadRequest:
      description: Dữ liệu đầu vào không hợp lệ
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
...
    MediaEvidence:
      type: object
      required: [mediaId, mimeType, sizeBytes, uploadedAt, publicUrl]
      properties:
        mediaId: { type: integer }
        mimeType: { type: string }
        sizeBytes: { type: integer }
        uploadedAt: { type: string, format: date-time }
        publicUrl: { type: string, format: uri }
    Error:
      type: object
      properties:
        error:
          type: string
        details:
          type: object
          additionalProperties: true
      required:
        - error
...
    AssetStatus:
      type: string
      enum:
        - AVAILABLE
        - RESERVED
        - BORROWED
        - DAMAGED
        - IN_REPAIR
        - RETIRED
...
    ReportIssueInput:
      type: object
      properties:
        description:
          type: string
          minLength: 1
          maxLength: 1000
      required:
        - description
      additionalProperties: false
...
    MediaEvidenceList:
      type: array
      items:
        $ref: '#/components/schemas/MediaEvidence'
...
    AssetIssueStatus:
      type: string
      enum:
        - REPORTED
        - CONFIRMED
        - REJECTED
        - CANCELLED
        - IN_REPAIR
        - COMPLETED
        - FAILED
...
    VendorReference:
      type: object
      description: Tên vendor được trả cho mọi user có asset_issue.view; contact fields chỉ có ở Vendor API.
      properties:
        id: { type: integer }
        name: { type: string }
      required: [id, name]
...
    AssetIssue:
      type: object
      properties:
        id:
          type: integer
        assetId:
          type: integer
        reportedBy:
          type: integer
          nullable: true
        description:
          type: string
          maxLength: 1000
        status:
          $ref: '#/components/schemas/AssetIssueStatus'
        createdAt:
          type: string
          format: date-time
          nullable: true
        updatedAt:
          type: string
          format: date-time
          nullable: true
        handledBy:
          type: object
          nullable: true
          description: Người xác nhận hoặc xử lý trạng thái gần nhất của issue.
          properties:
            id:
              type: integer
            name:
              type: string
          required:
            - id
            - name
        vendor:
          $ref: '#/components/schemas/VendorReference'
          nullable: true
        startDate:
          type: string
          format: date-time
          nullable: true
        endDate:
          type: string
          format: date-time
          nullable: true
        cost:
          type: string
          nullable: true
          description: Chi phí sửa chữa dạng chuỗi thập phân
        result:
          type: string
          maxLength: 300
          nullable: true
        note:
          type: string
          maxLength: 300
          nullable: true
        asset:
          type: object
          properties:
            id:
              type: integer
            serialNumber:
              type: string
              nullable: true
            status:
              $ref: '#/components/schemas/AssetStatus'
            modelName:
              type: string
          required:
            - id
            - status
            - modelName
          nullable: true
        reporter:
          type: object
          properties:
            id:
              type: integer
            name:
              type: string
          required:
            - id
            - name
          nullable: true
        repairEvidence:
          $ref: '#/components/schemas/MediaEvidenceList'
      required:
        - id
        - assetId
        - description
        - status
        - vendor
        - repairEvidence
...
security:
  - bearerAuth: []